requirements.txt
requests~=2.32.3
customtkinter~=5.2.2
pillow~=10.3.0
localStoragePy
aiohttp
pypresence
yt-dlp
CTkListbox